fixing a spot where the ai deemed bryan unintelligible
i had to slow it down to 50% speed to understand what he was saying but now there's a slightly more accurate transcription

by the way, not a transcription error but a content error/fyi for bryan & jess:

esr didn't write the jargon file. he just compiled a bunch of anecdotes. the only things he obviously wrote in it were political digs (for example the page where he proudly defiles a page to say programmers are right-wing because of the fall of the ussr). his book is an incredibly bad and not at all faithful rendition of the jargon file.

if you want to point people at a book next time this gets brought up, point it at guy steele's compilation of the jargon file. if you want to point them to the source, point them to the jargon file.